+++ src/main/java/io/kontur/eventapi/dao/mapper/DataLakeMapper.java
@@ -3,7 +3,6 @@ package io.kontur.eventapi.dao.mapper;
 import io.kontur.eventapi.entity.DataLake;
 import org.apache.ibatis.annotations.Mapper;
 
-import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;