+++ src/main/java/io/kontur/eventapi/entity/KonturEvent.java
@@ -5,15 +5,12 @@ import java.util.List;
 import java.util.UUID;
 
 public class KonturEvent {
-
     private UUID eventId;
-    private Long version;
     private List<UUID> observationIds = new ArrayList<>();
     private String provider;
 
-    public KonturEvent(UUID eventId, Long version) {
+    public KonturEvent(UUID eventId) {
         this.eventId = eventId;
-        this.version = version;
     }
 
     public UUID getEventId() {
@@ -32,14 +29,6 @@ public class KonturEvent {
         this.observationIds = observationIds;
     }
 
-    public Long getVersion() {
-        return version;
-    }
-
-    public void setVersion(Long version) {
-        this.version = version;
-    }
-
     public void addObservations(UUID observations) {
         this.observationIds.add(observations);
     }
@@ -56,7 +45,6 @@ public class KonturEvent {
     public String toString() {
         return "KonturEvent{" +
                 "eventId=" + eventId +
-                ", version=" + version +
                 ", observationIds=" + observationIds +
                 ", provider='" + provider + '\'' +
                 '}';