+++ src/main/java/io/kontur/eventapi/eventcombination/EventCombinator.java
@@ -2,12 +2,11 @@ package io.kontur.eventapi.eventcombination;
 
 import io.kontur.eventapi.entity.KonturEvent;
 import io.kontur.eventapi.entity.NormalizedObservation;
+import io.kontur.eventapi.job.Applicable;
 
 import java.util.Optional;
 
-public interface EventCombinator {
+public interface EventCombinator extends Applicable<NormalizedObservation> {
 
-    boolean isApplicable(NormalizedObservation normalizedObservation);
-
-    Optional<KonturEvent> findEventForObservation(NormalizedObservation normalizedObservation);
+    Optional<KonturEvent> findEventForObservation(NormalizedObservation observation);
 }